Save maps for inventories and guesses
This is so the inventory and guess columns,
which are often duplicated across session/player/team,
do not appear multiple times in the guesses data,
which makes the data pretty big.

I saw a reduction in size of the Guesses.rda by
removing those list columns from 30mb to 9mb.